pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "schema-manager"
version = "0.1.0"
description = "Schema lifecycle supervisor for sesam-opensource-poc"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30",
    "pyyaml>=6.0",
    "watchdog>=4.0",
    "aiohttp>=3.10",
    "prometheus-client>=0.21",
    "structlog>=24.0",
]

[project.scripts]
schema-manager = "schema_manager.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "testcontainers[postgres]>=4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
